requirements.txt
fastapi
uvicorn
python-dotenv
requests
pytest
passlib
bcrypt
python-multipart
stripe
itsdangerous
